requirements.txt
# Computer Vision (Contrib version for ChArUco calibration)
opencv-contrib-python==4.11.0.86

# Core dependencies
numpy==2.2.6

# Configuration management
PyYAML==6.0.2
